+++ package.json
@@ -12,13 +12,14 @@
   "dependencies": {
     "@expo-google-fonts/roboto": "^0.2.3",
     "@expo/vector-icons": "^14.0.4",
-    "@react-native-async-storage/async-storage": "^1.24.0",
-    "@react-native-firebase/firestore": "^21.6.1",
+    "@react-native-async-storage/async-storage": "1.23.1",
+    "@react-native-firebase/app": "^21.6.2",
+    "@react-native-firebase/firestore": "^21.6.2",
     "expo": "~51.0.28",
     "expo-checkbox": "~3.0.0",
     "expo-constants": "^16.0.2",
     "expo-font": "^12.0.10",
-    "expo-image-picker": "^16.0.3",
+    "expo-image-picker": "~15.0.7",
     "expo-linking": "^6.3.1",
     "expo-router": "^3.5.23",
     "expo-splash-screen": "^0.27.6",
@@ -28,14 +29,14 @@
     "react-dom": "18.2.0",
     "react-hook-form": "^7.54.0",
     "react-native": "0.74.5",
-    "react-native-safe-area-context": "^4.14.0",
-    "react-native-screens": "^4.0.0",
+    "react-native-safe-area-context": "4.10.5",
+    "react-native-screens": "3.31.1",
     "react-native-web": "~0.19.10"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
-    "@types/react": "~18.3.12",
-    "typescript": "^5.1.3"
+    "@types/react": "~18.2.79",
+    "typescript": "~5.3.3"
   },
   "private": true
 }
