+++ src/app/animalsList.tsx
@@ -1,23 +1,21 @@
 import { router } from 'expo-router';
-import { collection, getDocs } from 'firebase/firestore';
+import { collection, getDocs, getDoc, doc } from 'firebase/firestore';
 import React, { useEffect, useState } from 'react';
 import { ActivityIndicator, Image, ScrollView, StatusBar, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
 import Header from '../components/Header';
 import { db } from '../services/auth/firebase-config';
 
-
-
-const AnimalsInfoScreen = () => {
-
+const AnimalInfoScreen = () => {
   interface Pet {
     id: string;
     name: string;
-    photos: string;
+    photos?: string[];
     gender: string;
     age: string;
     size: string;
+    owner: string;
+    city: string;
   }
-
   const [pets, setPets] = useState<Pet[]>([]);
   const [loading, setLoading] = useState(true);
 
@@ -25,13 +23,28 @@ const AnimalsInfoScreen = () => {
     const fetchPets = async () => {
       try {
         const querySnapshot = await getDocs(collection(db, 'animals'));
-        const petsData = querySnapshot.docs.map((doc) => {
-          const data = doc.data() as Omit<Pet, 'id'>;
-          return {
-            id: doc.id,
-            ...data,
-          };
-        });
+
+        const petsData = await Promise.all(
+          querySnapshot.docs.map(async (docs) => {
+            const data = docs.data() as Omit<Pet, 'id' | 'localidade'>;
+            let city = 'Não informada';
+
+            if (data.owner) {
+              const ownerDoc = await getDoc(doc(db, 'users', data.owner));
+              if (ownerDoc.exists()) {
+                const ownerData = ownerDoc.data() as { city?: string };
+                city = ownerData.city || 'Não informada';
+              }
+            }
+
+            return {
+              id: docs.id,
+              ...data,
+              city,
+            };
+          })
+        );
+
         setPets(petsData);
       } catch (error) {
         console.error('Erro ao buscar os dados:', error);
@@ -47,9 +60,8 @@ const AnimalsInfoScreen = () => {
     return <ActivityIndicator size="large" color="#88c9bf" style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }} />;
   }
 
-
   return (
-    <ScrollView contentContainerStyle={styles.scroll}>
+    <ScrollView contentContainerStyle={{ flexGrow: 1 }}>
         <View style={styles.container}>
         <StatusBar backgroundColor={"#88c9bf"} barStyle={"light-content"} />
 
@@ -61,15 +73,16 @@ const AnimalsInfoScreen = () => {
           <View>
             <TouchableOpacity
               key={pet.id}
-              onPress={() => router.push(`./animal/${pet.id}`)} // Navega para a tela de detalhes
-              >
+              onPress={() => router.push(`./animal/${pet.id}`)}
+            >
             <Text style={styles.name}>{pet.name}</Text>
-            <Image source={{ uri: pet.photos[0] }} style={styles.image} />
-            <View style={{flexDirection:"row",justifyContent:"space-between",marginHorizontal:50}}>
+            <Image source={{ uri: pet.photos?.[0] }} style={styles.image} />
+          <View style={{ flexDirection: "row", justifyContent: "space-between", marginHorizontal: 50 }}>
             <Text style={styles.info}>{pet.gender}</Text>
             <Text style={styles.info}>{pet.age}</Text>
             <Text style={styles.info}>{pet.size}</Text>
-            </View>
+            <Text style={styles.info}>{pet.city}</Text>
+          </View>
             </TouchableOpacity>
           </View>
         </View>
@@ -81,15 +94,14 @@ const AnimalsInfoScreen = () => {
 };
 
 const styles = StyleSheet.create({
-    container: {
-        justifyContent: 'center',
-        alignItems: 'center',
-        position: 'relative',
-        backgroundColor: '#fafafa',
-      },
+  container: {
+    justifyContent: 'center',
+    alignItems: 'center',
+    position: 'relative',
+  },
+
   card: {
     backgroundColor: '#fff',
-    
     borderBottomLeftRadius: 8,
     borderBottomRightRadius: 8,
     borderTopLeftRadius: 3,
@@ -97,19 +109,20 @@ const styles = StyleSheet.create({
     marginBottom: 16,
     overflow: 'hidden',
     width: '90%',
-    elevation: 4, // Sombra para Android
-    shadowColor: '#000', // Sombra para iOS
+    elevation: 4,
+    shadowColor: '#000',
     shadowOffset: { width: 0, height: 2 },
     shadowOpacity: 0.2,
     shadowRadius: 4,
     top: 100,
   },
+
   image: {
     width: '100%',
     height: 150,
     resizeMode: 'cover',
   },
-  
+
   name: {
     backgroundColor: '#fee29b',
     fontSize: 18,
@@ -117,17 +130,11 @@ const styles = StyleSheet.create({
     color: '#333',
     padding: 5,
   },
+
   info: {
     fontSize: 14,
     color: '#555',
     marginBottom: 4,
   },
-  scroll: {
-    flex: 1,
-  },
 });
-
-
-
-export default AnimalsInfoScreen;
-
+export default AnimalInfoScreen;
