+++ src/app/index.tsx
@@ -4,21 +4,28 @@ import React from 'react';
 import { Image, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
 import Button from '../components/Button';
 import { auth } from '../services/auth/firebase-config';
+import { useSession } from '../services/auth/ctx';
 
 const Home = () => {
   const id = auth.currentUser?.uid;
   const router = useRouter();
+  const { user, isLoading } = useSession();
+
   return (
       <View style={styles.container}>
           <TouchableOpacity  style={styles.menuIcon} onPress={() => router.push(`./myAnimals/${id}`)}>
             <Feather name="menu" size={24} color="#88c9bf" />
           </TouchableOpacity>
         <Text style={styles.title}>Olá!</Text>
+        <StatusBar backgroundColor={"#88c9bf"} barStyle={"light-content"} />
+
+        <Feather style={styles.menuIcon} name="menu" size={24} color="#88c9bf" />
+        <Text style={styles.title}>Olá</Text>
         <Text style={[ styles.subtitle, { color: '#757575' }]}> Bem vindo ao Meau! Aqui você pode adotar, doar e ajudar cães e gatos com facilidade. Qual o seu interesse? </Text>
         
         <View style={styles.buttonContainer}>
           <Button title="ADOTAR" onPress={() => router.push('/animalsList')} variant="default" />
-          <Button title="AJUDAR" onPress={() => router.push('/ajudar')} variant="default" />
+          <Button title="AJUDAR" onPress={() => router.push('/login-screen-error')} variant="default" />
           <Button title="CADASTRAR ANIMAL" onPress={() => router.push('/registerAnimal')} variant="default" />
         </View>
